optional::transform: ensure that it works with non-movable payloads
This is supposed to work and it's explicitly spelled out in the standard
text.

If the function called by transform returns a non-movable prvalue,
we cannot rely on optional(T&&) constructor for this, because that will
not simply materialize it into the payload of the returned optional.

Instead, add a new private constructor for optional with a private tag.
Inside that constructor, we invoke the function and use it to initialize
the payload; this will not require any moves due to guaranteed copy
elision. In transform(), we call that constructor, passing the function
and the value stored in *this to it.

A similar change is needed for the optional<T&>.

Since we need to call this private optional<U>'s constructor from a
generic optional<T>::transform, we need to grant friendship to all
possible optional specializations.